pyproject.toml
[tool.poetry]
name = "pypeline-semantic-release"
version = "0.0.0"
description = "Pypeline step for semantic release."
authors = ["cuinixam <me@cuinixam.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/cuinixam/pypeline-semantic-release"
documentation = "https://pypeline-semantic-release.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "pypeline_semantic_release", from = "src" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/cuinixam/pypeline-semantic-release/issues"
"Changelog" = "https://github.com/cuinixam/pypeline-semantic-release/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.10"
py-app-dev = "^2.7"
typer = "^0.12"
pypeline-runner = "^1.6.0"
python-semantic-release = "^9.16.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^6.0"
pre-commit = "^4.1"
ruff = "^0.9"

[tool.poetry.group.docs.dependencies]
myst-parser = ">=0.16"
sphinx = ">=4.0"
sphinxcontrib-mermaid = "^0.8"
mlx-traceability = "^10.0"
sphinx-copybutton = "^0.5"
sphinx-new-tab-link = "^0.2"
sphinx-book-theme = "^1.1"
sphinx-design = "^0.5"

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variables = [
    "src/pypeline_semantic_release/__init__.py:__version__",
    "docs/conf.py:release",
]
build_command = "poetry build"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "chore*",
    "ci*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=pypeline_semantic_release --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.ruff]
target-version = "py38"
line-length = 180
lint.ignore = [
    "D203", # 1 blank line required before class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in `__init__`
    "D401", # First line of docstring should be in imperative mood
]
lint.select = [
    "B",   # flake8-bugbear
    "D",   # flake8-docstrings
    "C4",  # flake8-comprehensions
    "S",   # flake8-bandit
    "F",   # pyflake
    "E",   # pycodestyle
    "W",   # pycodestyle
    "UP",  # pyupgrade
    "I",   # isort
    "RUF", # ruff specific
]
exclude = [
    "bootstrap.py"
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "S101",
]
"setup.py" = ["D100"]
"conftest.py" = ["D100"]
"docs/conf.py" = ["D100"]

[tool.ruff.lint.isort]
known-first-party = ["pypeline_semantic_release", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
